+++ plugin.image.mypicsdb3/resources/lib/mypicsdb3/__init__.py
@@ -1,5 +1,5 @@
 """Core package for MyPicsDB 3."""
 
 ADDON_ID = "plugin.image.mypicsdb3"
-VERSION = "0.2.21"
-SCHEMA_VERSION = 3
+VERSION = "0.2.22"
+SCHEMA_VERSION = 4